requirements.txt
requests>=2.28.0
python-whois>=0.8.0
dnspython>=2.3.0
Pillow>=9.0.0
ipaddress 